Split command surface parity gates
